countCompletedMissions() 반환형 Long으로 변경

+++ src/main/java/com/umc/storeandmission/domain/mission/repository/MissionRepository.java
@@ -12,7 +12,7 @@ public interface MissionRepository extends JpaRepository<Mission, Long> {
             "join MemberMission mbm on m = mbm.mission " +
             "where mbm.member.memberId = :memberId " +
             "and m.store.region.regionId = :regionId ")
-    Integer countCompletedMissions(Long memberId, Long regionId);
+    Long countCompletedMissions(Long memberId, Long regionId);
 
     @Query("select m from Mission m " +
             "join MemberMission mbm on m = mbm.mission " +
